Allow the lid to be ignored using the use_lid pref
In the case where some devices have an optional lid,
the presence of the SW_LID in one of the input events
will cause powerm to treat that as a valid input and
suspend or shutdown the system appropriately.

Howevever, if the system doesn't physically have a lid
but has one logically, we want the power manager
to ignore inputs from that event based on a preference.
This is done at powerm initialization, and is not
dynamically changed.

This change requires this changelist :
http://codereview.chromium.org/6712051/

Change-Id: I16dba1389078a6dd367770db334f0aa8f4e20ec1
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Todd Broch <tbroch@chromium.org>

BUG=chromium-os:13260
TEST=
echo 0 > /var/lib/power_manager/use_lid
sudo restart powerm
Close the lid. Ensure that the lid is ignored.
Watch /var/log/power_manager/powerm.LATEST for "Ignoring Lid."
Boot a system without a lid.
Ensure the system doesn't shut down or suspend.

Review URL: http://codereview.chromium.org/6712052