Display off is done by a callback into backlight controller
Previously, backlight controller turned off the display after backlight
was turned off by using a polling event to wait for the backlight to
turn off.  But the polling loop caused problems.

This patch replaces the polling loop with a callback.  When the timeout
event that finally sets backlight to zero is triggered, it will invoke
the callback to determine whether the display should be turned off,
based on the power state.

BUG=chromium-os:13327
TEST=In /var/lib/power_manager, set [un]plugged_dim_ms to 30000 and
[un]plugged_off_ms to 60000.  Restart powerd.  Wait for 60 seconds.  The
display should turn off.

Change-Id: Ica4a1cc5708f28320607515e891f8aeee6831922
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6715014