Backlight off should not transition to dim or off states
When backlight has been set all the way down to zero, it should not
transition to the dim or off states due to idle.  Otherwise, user input
would cause the system to exit dim or idle and turn on the backlight.
It is more important to maintain the backlight off status.

BUG=chromium-os:13427
TEST=1. Set [un]plugged_dim_state and [un]plugged_off_state to 30000 in
/var/lib/power_manager, restart powerd.
2. Turn brightness down to zero, wait for more than 60 seconds.
3. Press a non-brightness key or move the touchpad.
4. The backlight should remain off.

Change-Id: I578f5024aa25d64a6dc2a6bf7fce51510729c640
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6733005