Minimum backlight brightness percent
Added minimum percentage brightness to backlight_controller.  This
is an optional power pref that is read from /usr/share/power_manager.
It defaults to zero.  This allows us to set a brightness minimum as
a percentage, based on LCD specs.

Also, replaces the brightness conversion formulas with a function
that uses the max brightness.  This makes the code cleaner -- easier to
read, and each conversion formula is defined only once.  The formula
also has the min brightness incorporated into it.  Currently min = 0, so
the full range of brightness levels is possible.  In the future, we can
add support for using an adjustable min brightness.

BUG=chromium-os:12103
TEST=Verify that manual brightness adjustment works over the whole
range.  Also try the "TEST=" in http://codereview.chromium.org/6733005
Also set min_backlight_percent to ~20 in /usr/share/power_manager,
and attempt to set brightness to zero.  It should not go below that level,
except go straight to zero.  Try manipulating the ALS to ensure that it
does not set brightness below the min.  Try idling and ensure that it
does not set brightness below the min.

Change-Id: I8ada74a7c9779d5920ef509ce299a9d475981f67
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6789006