Move backlight smoothing to backlight controller
Smooth, gradual transitions in backlight brightness has been implemented
in the backlight interface.  But in retrospect this was a poor design,
as it required a lot more extra code changes to run smoothly than it should.

This patch moves the smoothing mechanism to backlight controller.  This
is a better design --  it is more maintainable and keeps the code
simpler.

BUG=chromium-os:14074
TEST=With a kernel backlight driver that does not support gradual
transitions, adjust the brightness.  The transitions should appear
smooth.  This should work with both single key presses and holding down
the keys.

Change-Id: Ic66988e4c769d8ebe3639fc982c0d0e650a06c3b

Review URL: http://codereview.chromium.org/6813117