service,cryptohome: wire up lockbox to dbus
Lockbox provides install-time, tamper evident data.  This change wires it up to DBus.

BUG=chromium-os:13129
TEST=clear the tpm and reboot
          check /var/log/cryptohomed.log for Init message
          run cryptohome --action=status and check lockbox output
          run cryptohome --action=lockbox_test
          run cryptohome --action=status and check new fields exist
          reboot
          run cryptohome --action=status and ensure it is locked and has the data

I haven't yet figured out how to best integrate this into an autotest.  Basically, we need to save the TPM Owner password to disk and be able to tell cryptohomed to load it on demand to simulate a TpmClear event.

Review URL: http://codereview.chromium.org/6801020

Change-Id: I31c4bb4c739677ace9d995fae178bd9b3c52aa33