Adjust brightness at startup so that it's not too low
In an earlier patch, code was added to adjust brightness at startup so
that it's not below a certain minimum level.  But that was incorrectly
removed in a later patch.  This patch restores that change so that
brightness is not at zero at startup (it gets stuck there with recent
changes).

BUG=chromium-os:14201
TEST=Set brightness to zero and restart powerd.  Brightness should not
be stuck.

Change-Id: I18cc52e0d7a95c182362c4def2ad4e36098d55d5
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6882017