Make react time a prefs file instead of a constant
kReactMS=30000 <-- this is changed to a power constant that is read from
a file, just like the others in power_constants.cc.

BUG=chromium-os:11419
TEST=Set react_ms to something less than the default 30000.  Set
[un]plugged_dim_ms to something < 30000, but > react_ms.  Start powerd
and make sure it passes the CHECKs.  Now set the dim_ms to something <
react_ms.  Start powerd and make sure it fails the CHECKs.

Change-Id: I3d726e1fa3e8147e600c95f3c3d9c73a20c6188b
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6883064