Remove cached session_state_ from gobi_modem object.  Use seesion_id_ to gate disconnect.
This variable was doing us no good and it was causing a race between
when we returned from a Connect() DBus call (so that flimflam thought
it could try disconnecting us) and when SessionStateHandler got
called.

We now use session_id_ to gate disconnect, which seems substantially
more correct.

Change-Id: I7c92c9e657793b48d22eb898823dbd90ec3c0e96

R=ers@google.com,jglasgow@google.com
TEST=suite_Cellular
BUG=none

Review URL: https://chromereviews.googleplex.com/2811012