Backlight brightness percentage int64 -> double
(Revised after revert: updated unit tests.)

Backlight brightness as a percent was previously represented as a
64-bit integer type (int64).  This patch makes it represented as a
double-precision floating point (double).

The problem with using an integer percentage was the massive rounding
errors introduced when converting between percentage brightness and
actual backlight PWM brightness.  As a result, the brightness transition
step sizes became unpredictably nonuniform.  It always resulted in a
small remainder when the adjustment went from one end of the scale to
another.  Adjusting from 0% to 100% took a different path than adjusting
from 100% to 0%.  With floating point percentage, this is no longer a
problem.  Adjustment is the same both ways, assuming no changes due to
ambient light sensor reading.

Also, this patch changes the way the step size is calculated.  Instead
of being incremented by 1, to make sure it is nonzero, it is set to
the maximum of calculated step size and 1.

Also, this patch improves the transition logic between zero and nonzero,
especially in cases where min > 0.

BUG=chromium-os:14000
TEST=Type 'tail -f powerd.LATEST' to see the logging.  Adjust brightness
from 0% to 100%, and then from 100% to 0%.  Both times, there should be
no "leftover" at the end of the scale.  The steps should be exactly even
on the percentage scale.

Change-Id: If7d3e941d891a253b502640c4097f18c75750bc1
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6896015