Set kFuzzMS to be a pref file
Since we did this for kReactMS, we might as well do it for kFuzzMS.
It's a simple change and we might need to use it in the future, for the
same reason that we needed to make the change for kReactMS.

BUG=chromium-os:11419
TEST=In /var/lib/power_manager, set fuzz_ms to 60001 and restart powerd.
It should fail a check.  Remove it and it should run okay.

Change-Id: I7447b7397d87af59d7271e0a3b0c38499f1047a7
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6881092