Don't modprobe the TPM modules.
This is part of a cleanup of tpm/tcsd startup.  It also fixes ARM issues.

Our current strategy is to compile the TPM modules with the kernel, so
that the TPM is available earlier during boot.  A change by Will
(can't remember which one, sorry) took advantage of this to speed up
the boot.  We have done this for many months now.  We push image and
kernel changes together, so this will not break anything.

The driver's behavior is the same when the driver is linked in, since
the kernel command line sets the same arguments (force=1,
interrupts=0).

ARM kernel is still in flux, and currently, at least for some kernel
versions, the TPM driver is compiled as module.  As a result, the
tpm-probe code was trying to load the tpm_tis module (instead of
tpm_tis_i2c), causing weird errors.

Change-Id: Ib874bb2f23e42880d7ce9eed3f0f41db0f053e7c

BUG=chromium-os:14554
TEST=check that /var/log/messages does NOT contain "tpm is non-functional or tpm driver is missing" after boot

Committed: http://chrome-svn/viewvc/chromeos?view=rev&revision=e343bb6

Review URL: http://codereview.chromium.org/6883171