clean-shutdown: Emit a signal before doing anything.
This closes a race between flimflam and the shutdown process where flimflam can
see cromo's disconnect and try to do a reconnect, not knowing that cromo
disconnected because it was asked to terminate. We're guaranteed to see the dbus
signal indicating a shutdown starting before we see the dbus signal indicating a
disconnect.

BUG=chromium-os:9487
TEST=Adhoc
shutdown with 3g connected, and look for unfortunately-timed reconnects in
/var/log/messages.

Change-Id: Id7dff176c62959600b4690ff03198f20b2affebb
Signed-off-by: Elly Jones <ellyjones@chromium.org>