[login_manager] Clear exit handlers for session manager's children before shutdown
session_manager has handlers that watch for the exit of any managed
child process and attempt to restart it.  At shutdown, session_manager
forces all its children to exit.  These handlers should not fire,
because the next thing the session manager does after terminating its
children is exit.  It could be possible that the glib event loop would
get control in some instances between terming the children and exiting
the session_manager, and that it may choose to run one of the child
exit handlers.

To avoid this, we explicitly remove and destroy these handlers before
terminating all our children.  This means that there should be no way
that the session manager would attempt to restart chrome or anything
else upon exit.

BUG=chromium-os:14840
TEST=unit, install on device and run smoke_Suite

Change-Id: I6185d7b082f9765b339a8ec2ee73d64d1e661f09