GSM: Make Enable fail if the SIM card  is PIN-locked.
Even though Gobi allows the modem to be enabled when the SIM card is
PIN-locked, certain subsequent operations fail, such as getting the
IMSI and registering on a network. For consistency with the way other
modems behave, we now artificially cause Enable to fail when the
SIM is PIN-locked, so that the flimflam modem manager plugoin doesn't
have to handle two different cases, or delay reporting that the modem
is enabled.

BUG=chromium-os-partner:3538
TEST=Tested both talking directly to cromo with mm-enable, and via
flimflam with enable-device.

Change-Id: Ia56892da86c3f5f12ef103ac3a1f8150b3f1ffd2
Reviewed-on: https://gerrit-int.chromium.org/1156
Reviewed-by: Jason Glasgow <jglasgow@google.com>
Tested-by: Eric Shienbrood <ers@google.com>