Log brightness up/down events in metrics
This feature is used to keep track of whether users are adjusting
brightness up or down in battery and AC modes, which can suggest whether
the default brightness levels are too high or too low.

BUG=chromium-os:3324
TEST=Type: 'tail -f /var/log/power_manager/powerd.LATEST | grep "metric:
Brightness"'. Adjust the brightness up and down, on both AC and battery
power. There should be metrics traces for brightness up and down in
both AC and battery modes.

Change-Id: I959f56ba681ba00b3a181393ef3cea97c67f438c
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/651
Reviewed-by: Thiago Farina <tfarina@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>