init: enable support for per-user flimflam profiles
On login create the flimflam directory in cryptohome and push the user
profile. On logout pop the per-user profile. Handle the case of flimflam
crashing and being started by checking if we're logged in and explicitly
pushing on restart--this avoids doing a sleep of unknown duration after
starting flimflam to ensure it's ready for dbus cmds.

BUG=chromium-os:12948
TEST=manual:go to alt console; run list-profiles to verify only the default profile is present; login; check profiles again and verify the per-user profile is now present; logout and check again; also after kill flimflamd and verify profile stack is restored

Change-Id: I2253495df3362478c2c776434d846fee8793b7d9
Reviewed-on: http://gerrit.chromium.org/gerrit/1194
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>