gobi-cromo-plugin: Use short paths for devices.
There has been a bug in the device watcher for a long time that caused it to
access devices by short name ('qcqmi0') sometimes and long name ('/dev/qcqmi0')
other times. This caused spurious failures to delete devices and spurious starts
of the device watcher, since devices would not be seen to exist by control path
when they actually did exist.

Additionally, we can delete the invocation of RemoveDeviceByControlPath, since
we always call GetDeviceList before that point, and GetDeviceList will notice
immediately if a device is missing. Therefore, by the time we reach the
invocation of RemoveDeviceByControlPath, the device will already have been
detected to be missing and removed, so that removal is always spurious. Deleting
this invocationh also lets us delete RemoveDeviceByControlPath altogether.

BUG=chromium-os:15295
TEST=Adhoc
cromo -stderrthreshold 0 and watch debug output

Change-Id: Ida2a4f185ec503143b6335b652d3340daf2602c9
Signed-off-by: Elly Jones <ellyjones@google.com>
Reviewed-on: https://gerrit-int.chromium.org/1538
Reviewed-by: Jason Glasgow <jglasgow@google.com>