gobi-cromo-plugin: Undo udev hack partly.
I removed a call to RemoveDeviceByControlPath here, even though that call was
actually required. The udev event announcing that /dev/qcqmi0 is gone does not
mean that the sysfs file for the device is actually gone - the sysfs file
lingers as long as there are references to the device file held in kernelspace.

BUG=chromium-os:15295
TEST=Adhoc
modem_set_carrier a few times with 'modem status', checking carrier

Signed-off-by: Elly Jones <ellyjones@chromium.org>
Change-Id: I834eb166f7c4a50d65dbf81e3e0fa1361e8ac61e
Reviewed-on: https://gerrit-int.chromium.org/1619
Reviewed-by: Jason Glasgow <jglasgow@google.com>
Tested-by: Elly Jones <ellyjones@google.com>