Terminate any running cryptohome processes before cryptohomed stops
The cryptohome client may be in the middle of a long pkcs#11 initialization
operation with open files on the cryptohome mount. Since this is expected,
kill it in the cryptohomed pre-stop, so that the infraction is not logged by the
killers script.

BUG=chromium-os:15697
TEST=Log in to a fresh system, immediately shutdown (or reboot from a terminal
session). Restart and make sure that /var/log/shutdown_force_kill_processes is
empty.

Change-Id: I5d6fc9460891b4b82f810789ac5f8a42e307726d
Reviewed-on: http://gerrit.chromium.org/gerrit/1491
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>