[libchromeos] Take advantage of new params in logging callback, handle VLOG()
When we moved to the newer libbase stuff, the base::logging callbacks added a param called |message_start|
This points to the same place in the log message that we used to heurisitcally search for, so
now just go ahead and use it.

Also, VLOG() calls (which we can now do) pass a value for |severity| that we couldn't handle.  Now, set it to syslog's LOG_DEBUG level.

BUG=None
TEST=build shill with new libchromeos.a, run it and see verbose messages get logged correctly, along with normal LOG(INFO) stuff

Change-Id: Idcea88ba45add901685b94d157db788e9e2d5162
Reviewed-on: http://gerrit.chromium.org/gerrit/1510
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>