Make Pkcs#11 User Token checking more robust
Currently, we use some heuristics looking for the existence of certain files
to determine if PKCS#11 token needs initialization. This is very hacky and may
not catch all problem.h This CL changes IsTokenBroken() to also attempt opening
a session on the TPM slot and attempting to read basic information about the
token.

BUG=chromium-os:15694
TEST=manual
1) Attempt to log in and let token init complete, log in again - token init
   should be detected as having already succeeded.
2) Interrupt token init or try to break existing token. Log in again. Token
   should be correctly detected as being broken.
3) Blow away /home/chronos/user/.tpm, try pkcs11_init --pkcs11_token_status.
   Token status should show that token seems to be broken.

Change-Id: I2ab3d41503eab8e36d42abc7581fc111af46f3d2
Reviewed-on: http://gerrit.chromium.org/gerrit/1515
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Tested-by: Gaurav Shah <gauravsh@chromium.org>