Make login manager use crossystem to derive firmware information.
The chrome://system bios_info section reports information as
derived from the ACPI. The ACPI provided information changes as
the firmware development progresses. On top of that on the ARM
platforms the ACPI facility does not exist.

The portable way to derive the information system firmware and
boot status is the crossystem utility, which can rung on all
platforms and represents data consistently everywhere.

This CL make use of crossystem to gather the relevant firmware
information.

BUG=chromium-os:15800
TEST=manual

- build images for mario, alex and kaen
- install images on the appropriate platforms
- examine information under in chrome://system bios_info
- observe all platforms reporting reasonable information

Change-Id: Ic37bc58a7adfa43532bfeefd51fb38ecbab4792b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/1656
Reviewed-by: Randall Spangler <rspangler@chromium.org>