power_manager: Make powerd fetch session state at startup.
This makes powerd call the session manager's
RetrieveSessionState method at startup to determine whether
we're logged in or not and which user we're logged in as.

Previously, we'd always assume we weren't logged in, which
would make the power button shut down immediately.  This
was triggered whenever an update was applied, since that
would make us restart powerd.

BUG=chromium-os:15294
TEST=restarted powerd while logged in and checked that power button locks screen instead of shutting down; did the same while not logged in or logged in as guest and checked that we shut down immediately

Change-Id: I6be26cfa5f2e98c0e0147cc2d854433b9fc3ef3e
Reviewed-on: http://gerrit.chromium.org/gerrit/1732
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>