Honor users volume settings after logout -> login
the alsa-default.conf is made to present new users with reasonable
levels for various audio controls.  In the case of master volume,
chrome stores the user's preferences.

Previously, those preferences weren't being honored as there was a
race between the audio mixer in cros and alsa-default.conf  This
change lets the audio defaults get set prior to chrome evaluating and
setting audio preferences for the user.

BUG=chrome-os-partner:3613
TEST=manual,
1. Login to ChromeOS device.
2. goto http://html5zombo.com
3. Reduce the device volume to 10% (inaudible but not muted)
4. Logout, make sure volume is still 10% and login back.
5. Verify the video sound. (Device volume is still set to 10%).

Change-Id: I1440dbb76dfd5c2df1d2047bdcff6969dc70834f
Reviewed-on: http://gerrit.chromium.org/gerrit/1767
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>