Add in support for selecting WebUI Login by flag file
This adds a check for the existence of the file /root/.use_webui_login. If this file exists the session_manager_setup will add the --webui-login flag to the command line of chrome, which selects the WebUI login flow. It currently also sets the value of SKIP_OOBE to force skipping the OOBE. Once the WebUI OOBE is ready this will need to be removed in a follow on CL.

The makefile has been modified to touch use_webui_login on build and remove it on clean. This is needed by the ebuild to create the flag file in the image.

BUG=chromium-os:6398
TEST=Touched /root/.use_webui_login on image and WebUI login screen came up on boot.

Change-Id: If5d6714851e87fd2a2a6026d0444187ade71009b
Reviewed-on: http://gerrit.chromium.org/gerrit/1983
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>