ui: kill processes at logout in a more secure way
If we use kill -9 -1, we can kill processes atomically with respect to process
creation, so we just need to sudo to chronos and do it. We have to loop around
this process in case an attacker daemon is killing us as we appear before we
kill them, but the loop guarantees that the logout can't complete until we win.

BUG=chromium-os:8869
TEST=Adhoc
ssh to DUT, run some programs as chronos, then log out.

Change-Id: Iabbcee22a7c45c0529948bec5872c202f459cae0
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/1990
Reviewed-by: Will Drewry <wad@chromium.org>