Support the ModemManager State property, and send StateChanged signals for more state transitions.
A while ago, ModemManager introduced a State property on the Modem
interface. Until now, we were not assigning values to this property.
Instead, we were reporting state using a "state" entry in the
dictionary returned by the GetStatus method. This change eliminates
that non-standard behavior, and uses the State property instead.
It also sends the StateChange signal in more instances to better
emulate the behavior of ModemManager.

To support the Disabled -> Enabling -> Enabled transition and
the Enabled -> Disabling -> Disabled transition, we use a
PowerCallback handler to defer marking the modem as Enabled
or Disabled until the change has taken place and the callback
has been invoked.

BUG=none
TEST=The only real use we make in flimflam of the State property
(or the "state" item from GetStatus) is to detect a pre-existing
connection when flimflam starts up. I tested that this functionality
still works. Also, there are a limited number of connection state
transitions that flimflam acts on, and I tested that it still
reacts properly when those transitions occur.

Change-Id: I2721ab1dd10dd1b7531334bc72abc06f13b14c5f
Reviewed-on: https://gerrit-int.chromium.org/2579
Reviewed-by: Eric Shienbrood <ers@google.com>
Tested-by: Eric Shienbrood <ers@google.com>