vpn-manager: remove "tpm" from names, use "%usepeercert" to allow any server id
Since the slots and ids are pkcs#11 slots and IDs (that happen to use the PKCS#11 TPM token)
change the names to remove "tpm" from them.  Now that we've added the ability to connect
to any server with a proper CA signature in strongswan, pass %usepeercerts as rightid
whenever --server_id is not specified.

BUG=chromium-os:16279 chromium-os:16280
TEST=connect with certs from command line
connect with psk from command line

Change-Id: If03c69de8abb21c936ce55392d945530b728d8cd
Reviewed-on: http://gerrit.chromium.org/gerrit/2295
Tested-by: Ken Mixter <kmixter@chromium.org>
Reviewed-by: James Simonsen <simonjam@chromium.org>