Refactor the mounting code and improve unit test coverage.
In preparation for supporting FUSE-based file systems, this CL modifies
the disk manager to support mounting a device using the external mount
program as an alternative to the mount(2) library call.

The CL also refactors the disk manager code such that mount-related
functionalities are now handled by a mounter object.

BUG=none
TEST=Tested cros-disks on Cr48 by mounting different types of filesystem.

Change-Id: I865943d0ec7799c086a290b9b8d755337a399c6e
Reviewed-on: http://gerrit.chromium.org/gerrit/2970
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>