Make Enable fail if unable to get PIN status.
Cleaned up handling of failures getting PIN status, so that
Enable fails, and also so that the UnlockRetries property
doesn't get a nonsensical value.

BUG=chrome-os-partner:4141
TEST=With no SIM and the rfkill switch set to "on", try to enable
the modem, either using mm-enable or via flimflam. Before this
fix, cromo would crash from an assertion failure. Now, the enable
operation simply fails. I also retested normal cases in which a
SIM is present and the rfkill switch is "off", both when the SIM
is PIN-locked and when it is not.

Change-Id: Ideddc2e5140cdd16857152f326b7b36994c8be35
Reviewed-on: https://gerrit-int.chromium.org/2505
Tested-by: Eric Shienbrood <ers@google.com>
Reviewed-by: Jason Glasgow <jglasgow@google.com>