Replace /root/.dev_mode for crossystem call
In order to determine if we are booting a developer image we should call
crossystem cros_debug instead of checking for file /root/.dev_mode. This
change is related to recent change in the build_image script:
http://gerrit.chromium.org/gerrit/#change,3169

BUG=chromium-os:16648
TEST=Boot developer image and check that /usr/local is properly mounted.
Remove /root/.dev_mode and /mnt/stateful_partition/.developer_mode, reboot
the machine and the transition to developer mode should start (code falls
through if in line 149).

Change-Id: I3a88b842f81b8fbde98086d034e9b5cc54d92ba8
Reviewed-on: http://gerrit.chromium.org/gerrit/3399
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>