Treat Enable() as a async call because it now is
Treat Enable() as a async call because it now is, and ensure that
connect fails if it is called while the modem is in the process of
either enabling or disabling.

The background to this is that previously the DBus call to Disable
would succeed even though the call was in progress.  This allowed to
flimflam to immediately call Enable to re-enable the modem.  That call
to Enable would find that the modem was still enabled (because the
disable was not yet complete).  Enable would return successfully, and
milliseconds later the disable would complete and the modem would be
disabled.  Flimflam state was then inconsistent with cromo state.
Tests would fail and 3G was unusable.

BUG=chromium-os:17397
TEST=./run_remote_tests.sh --remote=jgcros1.cam --build network_3GSmoke

Change-Id: I03b91d7130c07c9192e28d0e30562e9b26c9e336
Reviewed-on: https://gerrit-int.chromium.org/3019
Reviewed-by: Jason Glasgow <jglasgow@google.com>
Tested-by: Jason Glasgow <jglasgow@google.com>
Reviewed-by: ttuttle <ttuttle@chromium.org>