Add DEVMODE check for mounting /usr/local
This change updates the DEVMODE variable to check for crossystem cros_debug,
which tells us if we have booted a developer image or an image in developer
mode. A base image that boots on developer mode now will mount /usr/local
correctly.

TEST=Run cros_au_test_harness. Also boot a base image in developer mode and
check for a correct mount of /usr/local
BUG=chromium-os:16648

Change-Id: I17b2399f03d5f53d18948a3406f4f5ae3aecc3ef
Reviewed-on: http://gerrit.chromium.org/gerrit/3952
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>