Brighten backlight when screen is locked
This patch addresses the state where the brightness is turned down to
zero and the screen is locked by idling. When the user returns, the
screen must be visible so the user knows what to do, i.e. type in
password and unlock.  The UI does not allow adjustment of brightness
when screen is locked, hence the screen must be brightened
automatically.

Also, manually locking screen when the brightness is zero should also
brighten the screen.

BUG=chromium-os:12347
TEST=Set the following variables in /var/lib/power_manager/:
echo 1 > lock_on_idle_suspend
echo 5000 | tee -a react_ms plugged_dim_ms unplugged_dim_ms
echo 10000 | tee -a plugged_off_ms unplugged_off_ms
echo 15000 > lock_ms
Type "restart powerd".  Set brightness to zero and wait 15 seconds for
it to lock the screen.  Press a key.  The brightness should come up
enough to show that the screen is locked.
Also, turn brightness to zero and hold the power button to lock.  The
brightness should immediately go up.

Change-Id: I5b140f61898203289235c1026429432441ee6b28
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4132
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>