Revert^2 "Repurpose the 11th partition for the read/write firmware."
This reverts commit 5fe6b43316ccaf0c32703043d10f933617956703 and also merges
it with a follow up change, originally described:

Adjust the position and size of the new R/W firmware partition.

Move it to start at 64KiB instead of immediately after the PMBR, GPT header,
and GPT table (17KiB). Also shrink it to be 8MiB instead of 9MiB-17KiB. The
bottom of the partition doesn't need to be aligned, and this preserves a spot
for the small, unused partitions it's displacing without forcing a second gap
to maintain alignment.

BUG=chrome-os-partner:4843
TEST=Build, boot and install an image. Verify the new size and position are
used and that the other partitions make sense.

Change-Id: Id5149f9ff8370da1945005ebeb3d3366fa409300
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4235
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>